Move Mp4Reader to reader.rs
